requirements.txt
# Core
numpy>=1.24,<2.0
scipy>=1.10
pandas>=2.0
h5py>=3.9
openpyxl>=3.1
pyyaml>=6.0
matplotlib>=3.7
tqdm>=4.66
einops>=0.7

# Deep learning (CUDA build selected by user)
torch>=2.1
tensorboard>=2.14
